+++ packages/ricos-converters/web/src/index.ts
@@ -7,7 +7,11 @@ import {
   toTiptapWithConverters as toTiptap,
   fromTiptapWithConverters as fromTiptap,
 } from './tiptap-converters';
+import { RichContent } from 'ricos-schema';
 export * from './types';
+
+const richContentNormalizer = (richContent: RichContent) =>
+  RichContent.toJSON(RichContent.fromJSON(richContent)) as RichContent;
 
 export const draftToTiptap: (draftContent: DraftContent) => JSONContent = flow(
   fromDraft,
@@ -15,5 +19,14 @@ export const draftToTiptap: (draftContent: DraftContent) => JSONContent = flow(
   JSON.parse,
   toTiptap
 );
-export const tiptapToDraft: (proseContent: JSONContent) => DraftContent = flow(fromTiptap, toDraft);
+
+export const tiptapToDraft = (
+  proseContent: JSONContent,
+  removeRichContentSchemaNormalizer = false
+): DraftContent => {
+  const richContent = fromTiptap(proseContent);
+  return toDraft(
+    removeRichContentSchemaNormalizer ? richContent : richContentNormalizer(richContent)
+  );
+};
 export { fromTiptap, toTiptap };